+++ plugins/types/custom_querier.go
@@ -2,11 +2,50 @@ package types
 
 import (
 	"encoding/json"
+	"errors"
+	bandwidthkeeper "github.com/cybercongress/go-cyber/v4/x/bandwidth/keeper"
+	dmnkeeper "github.com/cybercongress/go-cyber/v4/x/dmn/keeper"
+	graphkeeper "github.com/cybercongress/go-cyber/v4/x/graph/keeper"
+	gridkeeper "github.com/cybercongress/go-cyber/v4/x/grid/keeper"
+	rankkeeper "github.com/cybercongress/go-cyber/v4/x/rank/keeper"
 
 	errorsmod "cosmossdk.io/errors"
 	wasmvmtypes "github.com/CosmWasm/wasmvm/types"
 	sdk "github.com/cosmos/cosmos-sdk/types"
 )
+
+type ModuleQuerier interface {
+	HandleQuery(ctx sdk.Context, query CyberQuery) ([]byte, error)
+}
+
+var ErrHandleQuery = errors.New("error handle query")
+
+type QueryPlugin struct {
+	moduleQueriers []ModuleQuerier
+	rankKeeper     *rankkeeper.StateKeeper
+	graphKeeper    *graphkeeper.GraphKeeper
+	dmnKeeper      *dmnkeeper.Keeper
+	gridKeeper     *gridkeeper.Keeper
+	bandwidthMeter *bandwidthkeeper.BandwidthMeter
+}
+
+func NewQueryPlugin(
+	moduleQueriers []ModuleQuerier,
+	rank *rankkeeper.StateKeeper,
+	graph *graphkeeper.GraphKeeper,
+	dmn *dmnkeeper.Keeper,
+	grid *gridkeeper.Keeper,
+	bandwidth *bandwidthkeeper.BandwidthMeter,
+) *QueryPlugin {
+	return &QueryPlugin{
+		moduleQueriers: moduleQueriers,
+		rankKeeper:     rank,
+		graphKeeper:    graph,
+		dmnKeeper:      dmn,
+		gridKeeper:     grid,
+		bandwidthMeter: bandwidth,
+	}
+}
 
 func CustomQuerier(qp *QueryPlugin) func(ctx sdk.Context, request json.RawMessage) ([]byte, error) {
 	return func(ctx sdk.Context, request json.RawMessage) ([]byte, error) {
